+++ src/os/win/os_win.h
@@ -3,8 +3,8 @@
 
 _CAPI_BEGIN
 
-short* os_win_outputcoordget(short[2]);
-void os_win_outputcoordset(const short[2]);
+int* os_win_outputcoordget(int[2]);
+void os_win_outputcoordset(const int[2]);
 
 int os_win_outputgetc(void);
 int os_win_outputhasnline(void);
